LayoutTests/accessibility/mac/live-regions/live-region-types-expected.txt

This tests that the notifications for aria-live=polite and aria-live=assertive are properly formatted.

Received announcement request:
AnnouncementKey: Price: $10{
    AXLanguage = en;
}
AXPriorityKey: 10
AXAnnouncementIsLiveRegionKey: true

Received announcement request:
AnnouncementKey: Count: 5{
    AXLanguage = en;
}
AXPriorityKey: 90
AXAnnouncementIsLiveRegionKey: true


PASS successfullyParsed is true

TEST COMPLETE
Price: $10
Count: 5
